Removed deprecated persistence files
