--- data/testNursesFile
@@ -1,3 +1,0 @@
-Nursy,Nur,0100-0900,101
-Another,Nurse,0500-1300,102;103
-Last,One,0700-1500,0
