--- data/testOtherStaffFile
@@ -1,4 +1,0 @@
-Doctor,Docty,Doc,0130-0930
-Doctor,ER,Master,1500-2300
-Receptionist,Clerky,Clerk,0000-0800
-Receptionist,Head,Desk,0800-1600
