--- data/testPatientsFile
@@ -1,2 +1,0 @@
-Biggie,Cheese,20001014,12345,1130,1,Headache,None,Peanuts,Zofran,0
-Another,Patient,20090903,2666,1900,3,First seizure,None,None,Lotriderm,101
