+++ src/test/persistence/EncoderTest.java
@@ -1,34 +1,35 @@
-package deprecated.persistence;
+package persistence;
 
+import model.ActiveStaff;
+import model.CheckedInPatients;
+import model.exceptions.PatientCheckedInException;
+import model.exceptions.StaffClockedInException;
 import model.people.*;
 import model.rooms.Room;
+import org.json.simple.parser.ParseException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import persistence.Decoder;
+import persistence.Encoder;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
 
 /**
- * Tests for Writer class
+ * Tests for Encoder class
  */
 
-public class WriterTest {
+public class EncoderTest {
 
-    // From https://github.students.cs.ubc.ca/CPSC210/TellerApp
+    private static final String TEST_PATIENTS_FILE_PATH = "./data/testEncodePatientsFile.json";
+    private static final String TEST_NURSES_FILE_PATH = "./data/testEncodeNursesFile.json";
+    private static final String TEST_OTHER_STAFF_FILE_PATH = "./data/testEncodeOtherStaffFile.json";
 
-    private static final String TEST_PATIENT_FILE = "./data/testWritePatientsFile";
-    private static final String TEST_NURSES_FILE = "./data/testWriteNursesFile";
-    private static final String TEST_OTHER_STAFF_FILE = "./data/testWriteOtherStaffFile";
-
-    private Writer testWriterPatients;
-    private Writer testWriterNurses;
-    private Writer testWriterOtherStaff;
     private Patient testPatient1;
     private Patient testPatient2;
     private Nurse testNurse1;
@@ -40,17 +41,27 @@ public class WriterTest {
     private Room testRoom1;
     private Room testRoom2;
 
+    private CheckedInPatients testCheckedInPatients;
+    private ActiveStaff testActiveStaff;
+
     @BeforeEach
-    public void runBefore() throws FileNotFoundException, UnsupportedEncodingException {
-        testWriterPatients = new Writer(new File(TEST_PATIENT_FILE));
-        testWriterNurses = new Writer(new File(TEST_NURSES_FILE));
-        testWriterOtherStaff = new Writer(new File(TEST_OTHER_STAFF_FILE));
+    public void runBefore() {
+        CheckedInPatients.resetInstance();
+        testCheckedInPatients = CheckedInPatients.getInstance();
+        ActiveStaff.resetInstance();
+        testActiveStaff = ActiveStaff.getInstance();
 
         testPatient1 = new Patient("Bob", "Lee", 19700107, 12, 1430, 2,
                 "Pain everywhere", "Contact", "Nuts", "None");
         testPatient2 = new Patient("Jason", "Bourne", 19710415, 162, 1700,
                 3, "Amnesia", "None", "None", "Normality",
                 125);
+        try {
+            testCheckedInPatients.checkInPatient(testPatient1);
+            testCheckedInPatients.checkInPatient(testPatient2);
+        } catch (PatientCheckedInException e) {
+            fail("PatientCheckedInException was thrown.");
+        }
 
         ArrayList<Integer> roomsForTestNurse2 = new ArrayList<>();
         roomsForTestNurse2.add(115);
@@ -62,6 +73,16 @@ public class WriterTest {
         testReceptionist = new Receptionist("Charles", "Gregor", "0800-1600");
         testDoctor2 = new Doctor("Alena", "Klein", "1500-2300");
 
+        try {
+            testActiveStaff.clockIn(testNurse1, testNurse1.getShift());
+            testActiveStaff.clockIn(testDoctor1, testDoctor1.getShift());
+            testActiveStaff.clockIn(testReceptionist, testReceptionist.getShift());
+            testActiveStaff.clockIn(testDoctor2, testDoctor2.getShift());
+            testActiveStaff.clockIn(testNurse2, testNurse2.getShift());
+        } catch (StaffClockedInException e) {
+            fail("StaffClockedInException was thrown.");
+        }
+
         testRoom = new Room(125);
         testPatient2.assignRoom(testRoom);
 
@@ -72,13 +93,15 @@ public class WriterTest {
     }
 
     @Test
-    public void testWritePatients() {
-        testWriterPatients.write(testPatient1);
-        testWriterPatients.write(testPatient2);
-        testWriterPatients.close();
+    public void testEncodePatients() {
+        try {
+            Encoder.encodePatients(testCheckedInPatients.listOfCheckedInPatients(), TEST_PATIENTS_FILE_PATH);
+        } catch (IOException e) {
+            fail("An IOException was thrown when encoding the patient file.");
+        }
 
         try {
-            List<Patient> patients = Reader.readPatients(new File(TEST_PATIENT_FILE));
+            List<Patient> patients = Decoder.decodePatients(new File(TEST_PATIENTS_FILE_PATH));
 
             Patient patient1 = patients.get(0);
             assertEquals("Bob Lee", patient1.getFullName());
@@ -105,19 +128,22 @@ public class WriterTest {
             assertEquals(125, patient2.getRoomNumberToAssign());
 
         } catch (IOException e) {
-            fail("An unexpected IOException was thrown");
+            fail("An IOException was thrown when decoding the patient file.");
+        } catch (ParseException e) {
+            fail("A ParseException was thrown when decoding the patient file.");
         }
-
     }
 
     @Test
-    public void testWriteNurses() {
-        testWriterNurses.write(testNurse1);
-        testWriterNurses.write(testNurse2);
-        testWriterNurses.close();
+    public void testEncodeNurses() {
+        try {
+            Encoder.encodeNurses(testActiveStaff.getListOfActiveNurses(), TEST_NURSES_FILE_PATH);
+        } catch (IOException e) {
+            fail("An IOException was thrown when encoding the nurse file.");
+        }
 
         try {
-            List<Nurse> nurses = Reader.readNurses(new File(TEST_NURSES_FILE));
+            List<Nurse> nurses = Decoder.decodeNurses(new File(TEST_NURSES_FILE_PATH));
 
             Nurse nurse1 = nurses.get(0);
             assertEquals("Robert Stone", nurse1.getFullName());
@@ -134,19 +160,22 @@ public class WriterTest {
             assertEquals(117, nurse2.getRoomNumbersToAssign().get(1));
 
         } catch (IOException e) {
-            fail("An unexpected IOException was thrown");
+            fail("An IOException was thrown when decoding the nurse file.");
+        } catch (ParseException e) {
+            fail("A ParseException was thrown when decoding the nurse file.");
         }
     }
 
     @Test
-    public void testWriteOtherStaff() {
-        testWriterOtherStaff.write(testDoctor1);
-        testWriterOtherStaff.write(testReceptionist);
-        testWriterOtherStaff.write(testDoctor2);
-        testWriterOtherStaff.close();
+    public void testEncodeOtherStaff() {
+        try {
+            Encoder.encodeOtherStaff(testActiveStaff.getListOfActiveOtherStaff(), TEST_OTHER_STAFF_FILE_PATH);
+        } catch (IOException e) {
+            fail("An IOException was thrown when encoding the other staff file.");
+        }
 
         try {
-            List<Staff> otherStaff = Reader.readOtherStaff(new File(TEST_OTHER_STAFF_FILE));
+            List<Staff> otherStaff = Decoder.decodeOtherStaff(new File(TEST_OTHER_STAFF_FILE_PATH));
 
             Staff staff1 = otherStaff.get(0);
             assertEquals("Doctor", staff1.getPosition());
@@ -164,8 +193,9 @@ public class WriterTest {
             assertEquals("1500-2300", staff3.getShift());
 
         } catch (IOException e) {
-            fail("An unexpected IOException was thrown");
+            fail("An IOException was thrown when decoding the other staff file.");
+        } catch (ParseException e) {
+            fail("A ParseException was thrown when decoding the other staff file.");
         }
     }
-
 }
